[Backport] COMP: Update Qt install rules to install Qt5::Multimedia dependency
This commit ensures the MultimediaWidgets library is packaged in
Linux and Windows installers.

The MultimediaWidgets library is a dependency of the CTKWidgetsPythonQt
library built with multimedia wrapping following previous commit (ENH: Update
CTK to include QtMultimedia Python wrappers)

On macOS, this was not a problem because dependencies of all libraries
and plugins are automatically and recursively bundled.